Optmized zustand store
